requirements.txt
git+https://github.com/FaddyManatee/embed-pagination
git+https://github.com/FaddyManatee/nbsplayer
beautifulsoup4==4.11.1
discord.py==2.2.2  # then -> py -3 -m pip install -U discord.py[voice] or similar
python-dotenv==0.21.0
requests==2.28.1